+++ .env
@@ -1,7 +1,1 @@
-PGUSER=postgres
-PGHOST=192.168.191.2
-PGDATABASE=saKubo
-PGPASSWORD=sinag@123
-PGPORT=5432
-PORT=5000
-JWT_SECRET=t04TuKw87TZPArOpm35FyVQHfL3nJrbtIpoieCPCvDc=
+VITE_SAKUBO_APP_API_KEY=sakubo-1234567890abcdef1234567890abcdef
