--- src/api.js
@@ -1,8 +1,0 @@
-import axios from "axios";
-
-const api = axios.create({
-    baseURL: "http://sakubo-laravel-api.com", 
-    withCredentials: true,           
-});
-
-export default api;
